[hipfftw][helper and tests] Some refactoring and helper generalization in preparation of advanced and guru hipfftw plans (#2135)
## Motivations

- Main motivation: the upcoming developments for advanced and guru
hipfftw plans do require some amendments to hipfftw's implementation
details and, mostly, significant generalization steps for its test code
base. This PR intends to isolate the changes that are unrelated to the
new upcoming hipfftw features, in order to ease the upcoming code
reviews.
- Additional side motivations detailed below.

## Technical Details

- The next upcoming hipfftw features require `hipfftw_helper` to be
augmented with more `std::vector<ptrdiff_t>` members; some pre-existing
routines specific to its only such member thus far (`lengths`) are
generalized to that end. The structure's logic needs to be expanded for
non-default strides and/or (possibly multi-dimensional) batch sizes and
distances to be considered, along with safe type-conversion helper
routines;
- Given the specificities of hipfftw tests, being capable of reproducing
one given hipfftw (functional) test can be very helpful for verification
purposes and/or to easily compare with rocFFT (without the interfacing
layers). These changes enable a way to initialize such a
`hipfftw_functional_validation` test from a user-defined token (via the
added `hipfftw_token` test option);
- All messages for invalid-argument and runtime exceptions thrown by the
`hipfftw_helper` struct are revised for clarity;
- The changes to `hipfftw.cpp` mean to isolate data-layout-related
details as much as possible from the remaining logic, and to simplify
internal functions' signatures by expanding the
`hipfftw_general_layout_data` structure. Only `size_t` values are used
past user-facing entry points with these changes (the current limitation
for data layouts made of less than $2^{31}$ elements unless via `guru64`
APIs is actually inconsistent with FFTW3, hence removed herein);
- The motivations for the changes to `hipfftw_test.cpp` are twofold:
1) aligning the test logic with the expanded `hipfftw_helper` structure
and its corresponding members and member functions;
2) revising the parameter-generation steps for argument-validation tests
from an exhaustive brute-force strategy to probabilistic sampling
instead: the brute-force exhaustive strategy was found to produce an
excessive number of argument-validation tests for the (upcoming)
advanced and guru plan creation functions without significant test
benefits compared to a probabilistic sampling approach.
- The inclusion of `hip/hip_runtime_api.h` and/or `hip/hip_complex.h`
emits new warnings about ignored attributes (from clang) starting from
ROCm-7.0. These warnings are filtered out.

## Test Plan

These changes do not introduce new feature(s) and the current test base
suffices to cover the suggested changes. CI tests are expected to pass
reliably.

## Test Result

Successfully tested manually and by automated tests.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.
2

+++ clients/tests/multi_stream_test.cpp
@@ -140,7 +140,7 @@ public:
             std::vector<size_t> sub_strides(sub_dft.length.size());
             for(int stride_idx = sub_strides.size() - 1; stride_idx >= 0; stride_idx--)
             {
-                if(stride_idx == sub_strides.size() - 1)
+                if(static_cast<size_t>(stride_idx) == sub_strides.size() - 1)
                 {
                     sub_strides[stride_idx] = total_batch_for_step;
                 }
